pkg/prometheus: exclude status from hash computation
The operator computes a hash value which is derived from different
inputs such as the Prometheus object, the operator's configuration, the
StatefulSet spec, etc. The operator records the hash value as an
annotation on the generated StatefulSet then it compares the current and
new values to decide whether an update of the StatefulSet is needed or
not.

Because the operator updates the status subresource in parallel, the
status field of the Prometheus object should not be taken into account
to compute the hash value.

Signed-off-by: Simon Pasquier <spasquie@redhat.com>