pkg/prometheus: fix curl exec probe
By default, the `curl` command doesn't return with a non-zero exit code
if the endpoint doesn't reply with a 2xx status code. As a result, the
readiness probe would succeed as soon as the Prometheus web server is
ready to respond while `/-/ready` may return "503 Service Not Available"
(e.g. the server hasn't finished loading the TSDB). During a
StatefulSet update with a 2 (or more) replicas, it means that the
StatefulSet controller may recreate the next pod while the last updated
pod hasn't finished its initialization, leading to a service disruption.

The issue is fixed by adding the `--fail` flag to the command.

Note that the issue only affects Prometheus objects with
`spec.listenLocal: true`.

Signed-off-by: Simon Pasquier <spasquie@redhat.com>